+++ src/Query.php
@@ -11,6 +11,7 @@ use yii\base\Component;
 use yii\db\QueryInterface;
 use yii\db\QueryTrait;
 use Yii;
+use yii\helpers\ArrayHelper;
 
 /**
  * Query represents Mongo "find" operation.
@@ -340,12 +341,7 @@ class Query extends Component implements QueryInterface
         }
         $result = [];
         foreach ($rows as $row) {
-            if (is_string($this->indexBy)) {
-                $key = $row[$this->indexBy];
-            } else {
-                $key = call_user_func($this->indexBy, $row);
-            }
-            $result[$key] = $row;
+            $result[ArrayHelper::getValue($row, $this->indexBy)] = $row;
         }
         return $result;
     }