+++ src/file/ActiveQuery.php
@@ -178,6 +178,6 @@ class ActiveQuery extends Query implements ActiveQueryInterface
             }
         }
 
-        return $models;
+        return parent::populate($models);
     }
 }